+++ modules/cppql/include/cppql/expressions/comparison_expression.h
@@ -72,7 +72,10 @@ namespace sql
 
         [[nodiscard]] bool containsTables(const auto&... tables) const { return column.containsTables(tables...); }
 
-        void generateIndices(int32_t& idx) { index = idx++; }
+        void generateIndices(int32_t& idx)
+        {
+            if constexpr (!std::same_as<value_t, std::nullptr_t>) index = idx++;
+        }
 
         /**
          * \brief Generate expression comparing a column to a fixed or dynamic value.
@@ -80,31 +83,45 @@ namespace sql
          */
         [[nodiscard]] std::string toString() const
         {
+            if constexpr (std::same_as<value_t, std::nullptr_t>)
+            {
+                if constexpr (Op == ComparisonOperator::Eq)
+                    return std::format("{0} IS NULL", column.toString());
+                else
+                    return std::format("{0} IS NOT NULL", column.toString());
+            }
             // Format with column on LHS or RHS.
-            if constexpr (Lhs)
-                return std::format("{0} {1} ?{2}", column.toString(), ComparisonOperatorType<Op>::str, index + 1);
             else
-                return std::format("?{2} {1} {0}", column.toString(), ComparisonOperatorType<Op>::str, index + 1);
+            {
+                if constexpr (Lhs)
+                    return std::format("{0} {1} ?{2}", column.toString(), ComparisonOperatorType<Op>::str, index + 1);
+                else
+                    return std::format("?{2} {1} {0}", column.toString(), ComparisonOperatorType<Op>::str, index + 1);
+            }
         }
 
         void bind(Statement& stmt, const BindParameters bind) const
         {
-            if constexpr (std::is_pointer_v<value_t>)
+            if constexpr (std::same_as<value_t, std::nullptr_t>) {}
+            else if constexpr (std::is_pointer_v<value_t>)
             {
                 if (any(bind & BindParameters::Dynamic))
                 {
-                    if (value)
+                    if constexpr (!std::same_as<value_t, std::nullptr_t>)
                     {
-                        const auto res = stmt.bind(index + Statement::getFirstBindIndex(), *value);
-                        if (!res)
-                            throw SqliteError(
-                              std::format("Failed to bind dynamic parameter."), res.code, res.extendedCode);
-                    }
-                    else
-                    {
-                        if (const auto res = stmt.bind(index + Statement::getFirstBindIndex(), nullptr); !res)
-                            throw SqliteError(
-                              std::format("Failed to bind dynamic parameter."), res.code, res.extendedCode);
+                        if (value)
+                        {
+                            const auto res = stmt.bind(index + Statement::getFirstBindIndex(), *value);
+                            if (!res)
+                                throw SqliteError(
+                                  std::format("Failed to bind dynamic parameter."), res.code, res.extendedCode);
+                        }
+                        else
+                        {
+                            if (const auto res = stmt.bind(index + Statement::getFirstBindIndex(), nullptr); !res)
+                                throw SqliteError(
+                                  std::format("Failed to bind dynamic parameter."), res.code, res.extendedCode);
+                        }
                     }
                 }
             }
@@ -222,10 +239,8 @@ namespace sql
     template<is_result_expression C>
     [[nodiscard]] auto operator==(C&& col, std::nullptr_t)
     {
-        return ComparisonExpression<std::remove_cvref_t<C>,
-                                    typename std::remove_cvref_t<C>::value_t*,
-                                    ComparisonOperator::Eq,
-                                    true>(std::forward<C>(col), nullptr);
+        return ComparisonExpression<std::remove_cvref_t<C>, std::nullptr_t, ComparisonOperator::Eq, true>(
+          std::forward<C>(col), nullptr);
     }
 
     /**
@@ -237,10 +252,8 @@ namespace sql
     template<is_result_expression C>
     [[nodiscard]] auto operator==(std::nullptr_t, C&& col)
     {
-        return ComparisonExpression<std::remove_cvref_t<C>,
-                                    typename std::remove_cvref_t<C>::value_t*,
-                                    ComparisonOperator::Eq,
-                                    false>(std::forward<C>(col), nullptr);
+        return ComparisonExpression<std::remove_cvref_t<C>, std::nullptr_t, ComparisonOperator::Eq, false>(
+          std::forward<C>(col), nullptr);
     }
 
     ////////////////////////////////////////////////////////////////
@@ -316,10 +329,8 @@ namespace sql
     template<is_result_expression C>
     [[nodiscard]] auto operator!=(C&& col, std::nullptr_t)
     {
-        return ComparisonExpression<std::remove_cvref_t<C>,
-                                    typename std::remove_cvref_t<C>::value_t*,
-                                    ComparisonOperator::Ne,
-                                    true>(std::forward<C>(col), nullptr);
+        return ComparisonExpression<std::remove_cvref_t<C>, std::nullptr_t, ComparisonOperator::Ne, true>(
+          std::forward<C>(col), nullptr);
     }
 
     /**
@@ -331,10 +342,8 @@ namespace sql
     template<is_result_expression C>
     [[nodiscard]] auto operator!=(std::nullptr_t, C&& col)
     {
-        return ComparisonExpression<std::remove_cvref_t<C>,
-                                    typename std::remove_cvref_t<C>::value_t*,
-                                    ComparisonOperator::Ne,
-                                    false>(std::forward<C>(col), nullptr);
+        return ComparisonExpression<std::remove_cvref_t<C>, std::nullptr_t, ComparisonOperator::Ne, false>(
+          std::forward<C>(col), nullptr);
     }
 
     ////////////////////////////////////////////////////////////////