+++ version_history.md
@@ -4,6 +4,7 @@
 
 * Exposed error codes of the `sql::SqliteError` exception.
 * Fixed some operators not properly forwarding reference types.
+* Fixed `sql::ComparisonExpression` not properly handling `nullptr` binding.
 
 ## 0.1.0 - December 2022